Fix loop over segment functions
`get_func(seg->start_ea)` was used to get the first function in a
segment, but `get_func` actually returns the function whose *chunk* is
at some address. This means that if a segment starts with a function
chunk, its parent function's start EA will be used as starting point
for looping over functions in the segment, missing all functions before
it.